% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exCon-package.R
\docType{package}
\name{exCon-package}
\alias{exCon-package}
\title{Explore Contour Data Interactively}
\description{
exCon: Explore Contour Data Interactively
}
\details{
Interactive tools to explore topographic-like data
sets.  Such data sets take the form of a matrix in which the rows and
contain altitude/response information.  Such data is found in cartography,
2D spectroscopy and chemometrics.  The functions in this package create
interactive web pages showing the contoured data, possibly with
slices from the original matrix parallel to each dimension. The interactive
behavior is created using the D3.js 'JavaScript' library by Mike Bostock.
}
\author{
Bryan A. Hanson & Kristina R. Multry
}
\keyword{package}
