% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{exCon-package}
\alias{exCon-package}
\title{Lindenmayer System Functions}
\description{
exCon: Explore Contour Data Interactively
}
\details{
exCon is an interactive tool to explore topographic-like data
sets.  Such data sets take the form of a matrix in which the rows and
columns provide location/frequency information, and the matrix elements
contain altitude/response information.  Such data is found in cartography,
2D spectroscopy and chemometrics.  exCon creates an interactive web page
showing the contoured data set along with slices from the original matrix
parallel to each dimension. The page is written in d3/javascript.
}
\author{
Bryan A. Hanson & Kristina R. Multry
}
\keyword{package}

